// example/types/api.ts
export interface definitions {
// ...
  MergeClassRequest: {
    /**
     * 是否禁用
     */
    disable?: boolean;
    /**
     * 主键Id
     */
    id?: string;
    /**
     * 班级名称
     */
    name: string;
    /**
     * 备注
     */
    remarks?: string;
    /**
     * 管理老师账号，不传则不变
     */
    teacherCodes?: string;
  };
// ...
  Class: {
    /**
     * 禁用
     */
    disable?: boolean;
    /**
     * 主键ID
     */
    id?: string;
    /**
     * 班级名称
     */
    name?: string;
    /**
     * 备注
     */
    remarks?: string;
    /**
     * 学生人数 @Transient
     */
    stuNum?: number;
    /**
     * 老师ID @Transient
     */
    teacherCodes?: string;
    /**
     * 老师名称 @Transient
     */
    userNames?: string;
    /**
     * 是否是教师、管理员专用班级
     */
    vipClass?: boolean;
  };
// ...
}